backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.1
selenium>=4.15.0
Scrapy>=2.11.0
pandas>=2.1.1
python-dotenv>=1.0.1
openai>=1.60.1
google-api-python-client>=2.159.0
langchain>=0.3.15
langchain-community>=0.3.15
fake_useragent>=2.0.3
openpyxl>=3.1.5
asyncio>=3.4.3
aiohttp>=3.8.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
google-genai
google-generativeai
sentence-transformers